removal of TRA API column and reason why
Since we no longer send outcome data to the DQT, the API outcome error dashboard and declaration outcome section is broken.

This presents an issue as the current screen showing this is misleading and not reflective of the latest updates made to the service.
i.e. it still references sending outcomes to the TRA API, it not being recorded (when it actually is, just in a different place), and it still shows the big green Resend button.

 we probably need to update the designs on the declaration outcome section to remove the unnecessary fields and the big green resend button. I am not sure if this has already been covered in the new admin console designs.